+++ commands/copy.py
@@ -8,7 +8,7 @@ import click
 @click.option(
     '--no-cache',
     is_flag=True,
-    help='Copy directly from the source instead of the saved remake.')
+    help='Copy directly from the source instead of the saved recake.')
 def command(id: str = '', destination: str = '', no_cache: bool = False):
     """
     Copy a source to a destination.