+++ CHANGELOG.md
@@ -1,15 +1,18 @@
-# Remake Changelog
+# recake Changelog
 
 ## Pre-Releases:
+
+### v0.4.0
+ - Renamed the project.
 
 ### v0.3.0
  - `rm`, `open`, and `update` commands added.
  - `ls` changes.
+ - `add source` now just `add`.
 
 ### v0.2.0
  - `ls` command. 
  - Major refactor/restructure.  
 
-
 ### v0.1.0
-`add source`, `copy`, and `clip` commands.
+ - `add source`, `copy`, and `clip` commands.